run test/ts-decl-test.ts with mocha
